renamed to ros_ethercat. resolved a warning
