+++ CMakeLists.txt
@@ -1,6 +1,6 @@
 # http://ros.org/doc/groovy/api/catkin/html/user_guide/supposed.html
 cmake_minimum_required(VERSION 2.8.3)
-project(sr_ethercat)
+project(ros_ethercat)
 
 # Load catkin and all dependencies required for this package
 find_package(catkin REQUIRED COMPONENTS 
@@ -44,17 +44,17 @@ CATKIN_DEPENDS
 INCLUDE_DIRS
    include
 CFG_EXTRAS
-   sr_ethercat-extras.cmake.em
+   ros_ethercat-extras.cmake.em
 )
 
 add_definitions(-O3)
 
-add_executable(sr_ethercat src/main.cpp src/sr_ethercat.cpp)
-target_link_libraries(sr_ethercat rt ${catkin_LIBRARIES} ${Boost_LIBRARIES})
-add_dependencies(sr_ethercat ${catkin_EXPORTED_TARGETS})
-pr2_enable_rpath(sr_ethercat)
+add_executable(ros_ethercat src/main.cpp src/ros_ethercat.cpp)
+target_link_libraries(ros_ethercat rt ${catkin_LIBRARIES} ${Boost_LIBRARIES})
+add_dependencies(ros_ethercat ${catkin_EXPORTED_TARGETS})
+pr2_enable_rpath(ros_ethercat)
 
-install(TARGETS sr_ethercat
+install(TARGETS ros_ethercat
    DESTINATION ${CATKIN_PACKAGE_BIN_DESTINATION})
 
 file(GLOB PYTHON_SCRIPTS RELATIVE "${CMAKE_CURRENT_SOURCE_DIR}"
